+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Floating app bar on commentable item screen
 - Jump to relevant page & post when using "Open in browser"
 - Scroll highlighted comments into view
+- Faster image share
 
 ## v1.0.41